move getActiveEditorPosition to notebook instance
